fix pricing plans
