+++ packages/xy-shared/components/pricing-table/default-config.tsx
@@ -87,7 +87,7 @@ export const StarterPlan: SubscriptionPlan = {
     features.mitLicense,
   ],
   buttonLabel: 'Sign up',
-  buttonLabelSignedId: 'Subscribe',
+  buttonLabelSignedIn: 'Subscribe',
 };
 
 export const ProPlan: SubscriptionPlan = {
@@ -140,7 +140,7 @@ export const ProPlan: SubscriptionPlan = {
   ],
   highlighted: true,
   buttonLabel: 'Sign up',
-  buttonLabelSignedId: 'Subscribe',
+  buttonLabelSignedIn: 'Subscribe',
 };
 
 export const EnterprisePlan: SubscriptionPlan = {
@@ -182,7 +182,7 @@ export const EnterprisePlan: SubscriptionPlan = {
   ],
   buttonVariant: 'secondary',
   buttonLabel: 'Request a Quote',
-  buttonLabelSignedId: 'Request a Quote',
+  buttonLabelSignedIn: 'Request a Quote',
 };
 
 export const SubscriptionPlans: SubscriptionPlan[] = [
