+++ packages/xy-shared/components/pro/PricingTable/index.tsx
@@ -1,64 +1,7 @@
-'use client';
+import { requireSession } from '../../../lib/nhost';
+import PricingTableComponent from './pricing-table';
 
-import { useState } from 'react';
-import { callNhostFunction } from '../../../server-actions/call-nhost-function';
-import {
-  BillingInterval,
-  defaultSubscriptionPlans,
-  PricingTable,
-  SubscriptionPlanId,
-} from '../../pricing-table';
-
-export default function PricingTableComponent() {
-  const [plans, setPlans] = useState(defaultSubscriptionPlans);
-
-  const setLoading = (planId: SubscriptionPlanId, isLoading: boolean) => {
-    setPlans((plans) =>
-      plans.map((plan) =>
-        plan.id === planId
-          ? { ...plan, buttonLabel: isLoading ? 'Loading...' : 'Subscribe' }
-          : plan,
-      ),
-    );
-  };
-
-  const subscribe = async ({
-    plan,
-    interval,
-  }: {
-    plan: SubscriptionPlanId;
-    interval: BillingInterval;
-  }) => {
-    if (plan === SubscriptionPlanId.ENTERPRISE) {
-      return window.open('/pro/quote-request?plan=enterprise', '_blank');
-    }
-
-    setLoading(plan, true);
-
-    const response = await callNhostFunction('/stripe/create-checkout', {
-      plan,
-      interval,
-      framework: process.env.NEXT_PUBLIC_FRAMEWORK ?? 'react',
-    });
-
-    if (!response.error && response.url) {
-      window.location.href = response.url;
-      setTimeout(() => setLoading(plan, false), 500);
-      return;
-    }
-
-    setLoading(plan, false);
-  };
-
-  return (
-    <PricingTable
-      plans={plans}
-      onSelect={(subscribeConfig) =>
-        subscribe({
-          plan: subscribeConfig.plan,
-          interval: subscribeConfig.billingInterval,
-        })
-      }
-    />
-  );
+export default async function PricingTable() {
+  const { user } = await requireSession();
+  return <PricingTableComponent user={user} />;
 }
